Document Athanor-Server to Athanor interface
I can not find good documentation of the Athanor APIs but I tried to
document this interace as much as I understand this far.
I also added documentation of how an Athanor-Server developer
can point at the versioned distribution Athanor library jars that
we are now building, or the latest copy of the Athanor library
jar for testing.